+++ src/quest/manager.py
@@ -133,9 +133,6 @@ class WorkflowManager:
     def get_wrapped_resource_stream(self, workflow_id: str, identity):
         return self._get_workflow(workflow_id).get_wrapped_resource_stream(identity)
 
-    async def wait_for_completion(self, workflow_id: str, identity):
-        return await self._get_workflow(workflow_id).wait_for_completion(identity)
-
     async def send_event(self, workflow_id: str, name: str, identity, action, *args, **kwargs):
         return await self._get_workflow(workflow_id).record_external_event(name, identity, action, *args, **kwargs)
